Move hand, sort reverse
