+++ tower-defense/src/game_state/mod.rs
@@ -7,7 +7,6 @@ pub mod fast_forward;
 mod field_area_effect;
 pub mod field_particle;
 pub mod flow;
-pub mod hand;
 pub mod item;
 mod level_rarity_weight;
 mod modal;
@@ -26,6 +25,7 @@ mod tower_info_popup;
 pub mod upgrade;
 mod user_status_effect;
 
+use crate::hand::HandSlotId;
 use crate::route::*;
 use crate::*;
 use background::{Background, generate_backgrounds};
@@ -34,7 +34,6 @@ use cursor_preview::CursorPreview;
 use fast_forward::FastForwardMultiplier;
 use field_area_effect::FieldAreaEffect;
 use flow::GameFlow;
-use hand::HandSlotId;
 use item::Item;
 pub use level_rarity_weight::level_rarity_weight;
 pub use modal::Modal;
