+++ tower-defense/src/hand/render_tower.rs
@@ -1,4 +1,4 @@
-use super::shared::{render_background_rect, render_top_left_rank_and_suit};
+use super::*;
 use crate::{
     asset_loader::get_tower_asset,
     game_state::tower::{AnimationKind, TowerKind, TowerTemplate},
