+++ tower-defense/src/hand/shared.rs
